out controller

+++ main.py
@@ -2,7 +2,7 @@ from pyrogram import Client, filters
 
 from time import sleep
 
-app = Client("account_1")
+app = Client("farm_account")
 
 CanWork = False
 IsWorking = False
@@ -10,32 +10,45 @@ IsWorking = False
 CanSendKits = False
 IsSending = False
 
-@app.on_message(filters.text & filters.me & filters.command("status", prefixes="."))
+def CheckUser(message):
+    ReoliedUser = app.get_messages(message.chat.id, reply_to_message_ids=message.message_id).from_user
+    if ReoliedUser.is_self:
+        return True
+    else:
+        return False
+
+@app.on_message(filters.text & filters.command("status", prefixes="."))
 def Status(client, message):
+    if not CheckUser(message):
+        return None
     message.delete()
-    message.reply_text(f"IsEating = {IsSending}\n"
-                       f"IsWorking = {IsWorking}")
+    message.reply_text(f"IsWorking = {IsWorking}\n"
+                       f"IsSending = {IsSending}")
 
 
-@app.on_message(filters.text & filters.me & filters.command("settings", prefixes="."))
+@app.on_message(filters.text & filters.command("settings", prefixes="."))
 def Settings(client, message):
+    if not CheckUser(message):
+        return None
     message.delete()
     message.reply_text(f"CanWork = {CanWork}\n"
                        f"CanSendKits = {CanSendKits}")
 
 
-@app.on_message(filters.command("send", prefixes=".") & filters.me)
+@app.on_message(filters.command("send", prefixes="."))
 def SendKitsCommand(client, message):
     command = message.text.split(".send ", maxsplit=1)[1]
-    message.delete()
+    if not CheckUser(message):
+        return None
+    #message.delete()
 
     global CanSendKits
     global IsSending
 
     if command == "start":
         CanSendKits = True
-        SendKits(message)
         message.reply_text("Send started")
+        SendKits(message)
     elif command == "stop":
         CanSendKits = False
         IsSending = False
@@ -46,21 +59,23 @@ def SendKitsCommand(client, message):
 def SendKits(message):
     global IsSending
 
-    Receiver = app.get_messages(message.chat.id, reply_to_message_ids=message.message_id)
+    #Receiver = app.get_messages(message.chat.id, reply_to_message_ids=message.message_id)
     if not IsSending and CanSendKits:
         IsSending = True
         while IsSending:
-            message.send_message(Receiver.chat.id, "Отправить аптечки 10", reply_to_message_id=Receiver.message_id)
+            app.send_message(message.chat.id, "Отправить аптечки 10", reply_to_message_id=message.message_id)
             sleep(86410) #86400
         IsSending = False
     else:
         message.reply_text("Is sending")
 
 
-@app.on_message(filters.command("work", prefixes=".") & filters.me)
+@app.on_message(filters.command("work", prefixes="."))
 def WorkCommand(client, message):
+    if not CheckUser(message):
+        return None
     command = message.text.split(".work ", maxsplit=1)[1]
-    message.delete()
+    #message.delete()
 
     global CanWork
     global IsWorking
@@ -93,3 +108,5 @@ def Working(message, work):
         message.reply_text("Работа завершена", quote=False)
     else:
         message.reply_text("Is Working")
+
+app.run()
